🚚 Moved components directory to root
