+++ components/aboutMeBanner/aboutMeBanner.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles.scss';
+@import '/styles/styles.scss';
 
 .aboutMeSection {
     background-color: $background-color;
