+++ components/myNavbar/myNavbar.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles.scss';
+@import '/styles/styles.scss';
 
 .myNavbar {
     background-color: $background-color !important;
