+++ components/socialsIcons/socialsIcons.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles.scss';
+@import '/styles/styles.scss';
 
 .socialsIconsContainer {
     align-content: center;
