+++ components/techIcon/TechIcon.tsx
