+++ components/typeBadge/TypeBadge.tsx
