+++ components/typeBadge/typeBadge.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles.scss';
+@import '/styles/styles.scss';
 
 .typeBadge {
     color: $color-white !important;
