+++ components/workExperienceSection/workExperienceSection.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles.scss';
+@import '/styles/styles.scss';
 
 .sectionTitle {
     @include section-title-text;
