+++ components/workPolaroid/workPolaroid.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/styles.scss';
+@import '/styles/styles.scss';
 
 .workPolaroidCard {
     @include animated-box-shadow;
