+++ pages/_home/myLinks/MyLinks.tsx
@@ -2,7 +2,7 @@ import { Button, Container } from 'react-bootstrap';
 import styles from './myLinks.module.scss';
 import { openLink } from '@/lib/utils';
 import ILinkUrl from 'core/interfaces/linkUrl';
-import SocialIconImage, { IconSize } from '../../../core/components/socialsIcons/SocialIconImage';
+import SocialIconImage, { IconSize } from 'components/socialsIcons/SocialIconImage';
 import { linkUrls } from 'core/data/linkUrls';
 
 function MyLinks() {
