+++ pages/experience/index.tsx
@@ -1,12 +1,12 @@
 import styles from './experience.module.scss';
 import IExperience from 'core/interfaces/experience';
-import TechIcon from 'core/components/techIcon/TechIcon';
+import TechIcon from 'components/techIcon/TechIcon';
 import { getGitConnectedPortfolio } from 'core/lib/api';
 import { useEffect, useState } from 'react';
 import { getMonthName, getTechIconsByJobName } from '@/lib/utils';
 import Accordion from 'react-bootstrap/Accordion';
 import IGitConnectedWork from 'core/interfaces/gitConnectedWork';
-import PageContainer from 'core/components/containers/PageContainer';
+import PageContainer from 'components/containers/PageContainer';
 
 type Props = {
     experiences: IGitConnectedWork[]
