+++ pages/index.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from "react";
 import Header from "./_home/header/Header";
-import MyNavbar from 'core/components/myNavbar/MyNavbar';
+import MyNavbar from 'components/myNavbar/MyNavbar';
 import MyLinks from "./_home/myLinks/MyLinks";
 import LoadingIcon from "pages/_home/loadingIcon/LoadingIcon";
 import MyWork from "./_home/myWork/MyWork";
-import AboutMeBanner from "../core/components/aboutMeBanner/AboutMeBanner";
-import WorkExperienceSection from "@/components/workExperienceSection/workExperienceSection";
+import AboutMeBanner from "components/aboutMeBanner/AboutMeBanner";
+import WorkExperienceSection from "components/workExperienceSection/workExperienceSection";
 
 function App() {
   const [fetchingData, setFetchingData] = useState(true);
