+++ pages/work/index.tsx
@@ -1,8 +1,8 @@
 import styles from './work.module.scss';
 import { projects } from 'core/data/projects';
-import WorkPolaroid from 'core/components/workPolaroid/WorkPolaroid';
+import WorkPolaroid from 'components/workPolaroid/WorkPolaroid';
 import { IWorkProject } from 'core/interfaces/workProject';
-import PageContainer from 'core/components/containers/PageContainer';
+import PageContainer from 'components/containers/PageContainer';
 
 type Props = {
     workProjects: IWorkProject[]
